requirements.txt
aenum==3.1.5
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///home/builder/ci_310/cffi_1642753365720/work
charset-normalizer==2.0.12
cryptography @ file:///home/builder/ci_310/cryptography_1640808721005/work
docopt==0.6.2
docutils==0.18.1
ffpyplayer==4.3.2
idna @ file:///tmp/build/80754af9/idna_1622654382723/work
iris==1.0.7
Kivy==2.1.0
Kivy-Garden==0.1.4
kivy-garden.mapview==1.0.5
mapview==1.0.5
msgpack==1.0.3
packaging==21.3
pipreqs==0.4.11
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.11.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.7
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-can==4.0.0
requests==2.27.1
typing_extensions==4.1.1
urllib3==1.26.7
wrapt==1.13.3
yarg==0.1.9
